feat(core): Baseclass for agents using websockets
Formatting is fixed as well in the legacy UI
