+++ core/agents/neo_think_agent.py
@@ -2,6 +2,8 @@ from typing import Annotated
 from typing_extensions import TypedDict
 import os
 
+from fastapi.websockets import WebSocket
+
 from langgraph.graph.message import add_messages
 from langchain_openai import ChatOpenAI
 from langchain_core.tools import tool
@@ -9,44 +11,36 @@ from langgraph.checkpoint.memory import MemorySaver
 from langgraph.graph import MessagesState, StateGraph, START, END
 from langchain_core.messages import BaseMessage, AIMessageChunk, HumanMessage, AIMessage, ToolMessage
 from langgraph.prebuilt import ToolNode, tools_condition
+
 from agents.model import Model #Models for chatGPT
 
 # Premade tool imports
 from langchain_community.tools.tavily_search import TavilySearchResults
 # Custom tool imports
 from tools.add_tool import add # Adds 2 numbers together
+from tools.youtube_transcript import youtube_transcript
 
-from jarvis.core.chroma import init_chroma, create_collection, add_document, upsert_document
+from agents.WebSocketAgent import WebSocketAgent # Superclass
 
-from agents.WebSocketAgent import WebSocketAgent
-"""
-Neoagent uses the ReAct agent framework.
-Simply put in steps:
-1. 'Re' The agent reasons about the problem, and plans out steps to solve it.
-2. 'Act' The agent acts upon the information gathered. Calling tools or interacting with systems based on the earlier reasoning.
-3. 'Loop' If the problem is not adequately solved, the agent can reason and act recursively until a satisfying solution is reached.
-ReAct is a simple multi-step agent architecture.
-Smaller graphs are often better understood by the LLMs.
-"""
-class NeoAgent(WebSocketAgent):
+class NeoThinkAgent(WebSocketAgent):
     def __init__(self):
         print("Instantiated NeoAgent....")
         system_prompt = "You are Jarvis, an AI assistant here to help the human accomplish tasks. Respond in a conversational, natural style that sounds good when spoken aloud. Keep responses short and to the point, using clear, engaging language. When explaining your thought process, be concise and only describe essential steps to maintain a conversational flow."
         # Defining the model TODO: Make this configurable with Llama, Grok, Gemini, Claude
         model = ChatOpenAI(
-            model = Model.gpt_4o_mini,
+            model = Model.gpt_4o_mini[0],
             temperature=0,
             max_tokens=16384, # Max tokens for mini. For gpt4o it's 128k
         ) # Using ChatGPT hardcoded (TODO: Make this dynamic)
         # Defining the checkpoint memory saver.
         memory = MemorySaver()
         # Tools list
-        tools = [add]
+        tools = [add, youtube_transcript]
 
         if os.getenv("TAVILY_API_KEY"):
             # Defining the tavily web-search tool
             tavily = TavilySearchResults(max_results=2)
-            tools = [add, tavily]
+            tools.append(tavily)
         else:
             print("TAVILY_API_KEY does not exist.")
             
@@ -78,8 +72,8 @@ class NeoAgent(WebSocketAgent):
         self.graph = graph_builder.compile(checkpointer=memory)
 
         # Draws the graph visually
-        with open("neoagent.png", 'wb') as f:
-            f.write(self.graph.get_graph().draw_mermaid_png())
+        # with open("neoagent.png", 'wb') as f:
+          #  f.write(self.graph.get_graph().draw_mermaid_png())
 
     # Streams graph updates using websockets.
     def stream_graph_updates(self, user_input: str):
@@ -87,8 +81,3 @@ class NeoAgent(WebSocketAgent):
         for event in self.graph.stream({"messages": [("user", user_input)]}, config):
             for value in event.values():
                 print("Assistant:", value["messages"][-1].content)
-
-    #Initialize chroma client and creates a collection when jarvis start
-    async def initialize_jarvis(self):
-        self.chroma_client = await init_chroma()
-        self.collection = await create_collection("my_collection", self.chroma_client)
